deepfake-detector/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
facenet-pytorch>=2.5.3
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
kagglehub>=0.2.0
gdown>=4.7.0
